+++ src/locale/en-us.js
@@ -10,15 +10,26 @@ const numbers = {
 	romanOrdinalIndicator :		"",
 }
 
-const abbreviationsForNbsp = [
-	"e.g.",
-	"i.e.",
-	"p.",
-	"pp."
+const singleWordAbbreviations = [
+	"p",
+	"pp",
+	"no",
+	"vol"
+]
+
+// Disregard correct spelling in your locale
+// and provide abbreviations in format “abbr abbr abbr…”
+const multipleWordAbbreviations = [
+	"U S",
+	"e g",
+	"i e",
+	"a m",
+	"p m",
 ]
 
 export default {
 	quotes,
 	numbers,
-	abbreviationsForNbsp
+	singleWordAbbreviations,
+	multipleWordAbbreviations
 }