+++ src/locale/rue.js
@@ -10,10 +10,16 @@ const numbers = {
 	romanOrdinalIndicator :		"\\.",
 }
 
-const abbreviationsForNbsp = []
+const singleWordAbbreviations = []
+
+// Disregard correct spelling in your locale
+// and provide abbreviations in format “abbr abbr abbr…”
+const multipleWordAbbreviations = [
+]
 
 export default {
 	quotes,
 	numbers,
-	abbreviationsForNbsp
+	singleWordAbbreviations,
+	multipleWordAbbreviations
 }