+++ src/locale/sk.js
@@ -7,29 +7,43 @@ const quotes = {
 
 const numbers = {
 	ordinalIndicator :				"\\.",
-	romanOrdinalIndicator :				"\\.",
+	romanOrdinalIndicator :		"\\.",
 }
 
-const abbreviationsForNbsp = [
-	"č.",
-	"s.",
-	"fol.",
-	"str.",
-	"r.",
-	"par.",
-	"odst.",
-	"napr.",
-	"sv.",
-	"tj.",
-	"tzv.",
-	"čl.",
-	"hl.",
-	"m."
+const singleWordAbbreviations = [
+	"č",
+	"s",
+	"fol",
+	"str",
+	"r",
+	"par",
+	"odst",
+	"napr",
+	"sv",
+	"tzv",
+	"čl",
+	"cit",
+	"roč",
+	"vyd"
 ]
 
+// Disregard correct spelling in your locale
+// and provide abbreviations in format “abbr abbr abbr…”
+const multipleWordAbbreviations = [
+	"hl m",
+	"n l",
+	"p n l",
+	"pr n l",
+	"s a",
+	"s l",
+	"t j",
+	"zodp red",
+	"t č"
+]
 
 export default {
 	quotes,
 	numbers,
-	abbreviationsForNbsp
+	singleWordAbbreviations,
+	multipleWordAbbreviations
 }