+++ src/test/punctuation/dash.test.js
@@ -1,8 +1,7 @@
 import {replaceThreeHyphensWithEmDash,
-				replaceTwoHyphensWithEnDash,
-				replaceSpacedHyphenWithEmDash,
-				replaceSpacedEnDashWithEmDash,
-				fixSpacesAroundEmDash,
+				replaceSpacedHyphenWithDash,
+				consolidateSpacedDashes,
+				fixDashSpacesBetweenWords,
 				fixDashBetweenCardinalNumbers,
 				fixDashBetweenOrdinalNumbers,
 				fixDash} from "../../lib/punctuation/dash";
@@ -21,60 +20,73 @@ describe('Replace 3 hyphens with an em dash\n', () => {
 	});
 });
 
-describe('Replace 2 hyphens with an en dash\n', () => {
-	let testCase = {
-		"Brno--Praha": "Brno–Praha",
-		};
-
-	Object.keys(testCase).forEach((key) => {
-		it("unit test", () => {
-			assert.equal(replaceTwoHyphensWithEnDash(key, new Locale("en-us")), testCase[key]);
-		});
-		it("module test", () => {
-			assert.equal(fixDash(key, new Locale("en-us")), testCase[key]);
-		});
-	});
-});
-
-describe('Replace spaced hyphen with an em dash\n', () => {
+describe('Replace spaced hyphen with an em dash (en-us, sk, cs, rue)\n', () => {
 	let testCase = {
 		"and - she said": "and — she said",
 		};
 
 	Object.keys(testCase).forEach((key) => {
 		it("unit test", () => {
-			assert.equal(replaceSpacedHyphenWithEmDash(key, new Locale("en-us")), testCase[key]);
+			assert.equal(replaceSpacedHyphenWithDash(key, new Locale("en-us")), testCase[key]);
+			assert.equal(replaceSpacedHyphenWithDash(key, new Locale("cs")), testCase[key]);
 		});
 	});
 });
 
-describe('Replace spaced en dash with an em dash\n', () => {
+describe('Replace spaced hyphen with an en dash (de-de)\n', () => {
+	let testCase = {
+		"und - er sagte": "und – er sagte",
+		};
+
+	Object.keys(testCase).forEach((key) => {
+		it("unit test", () => {
+			assert.equal(replaceSpacedHyphenWithDash(key, new Locale("de-de")), testCase[key]);
+		});
+	});
+});
+
+describe('Replace spaced en dash with an em dash (en-us, sk, cs, rue)\n', () => {
 	let testCase = {
 		"and – she said": "and — she said",
 		};
 
 	Object.keys(testCase).forEach((key) => {
 		it("unit test", () => {
-			assert.equal(replaceSpacedEnDashWithEmDash(key, new Locale("en-us")), testCase[key]);
+			assert.equal(consolidateSpacedDashes(key, new Locale("en-us")), testCase[key]);
+			assert.equal(consolidateSpacedDashes(key, new Locale("sk")), testCase[key]);
 		});
 	});
 });
 
-describe('Fix spaces around em dash in English\n', () => {
+describe('Replace spaced em dash with an en dash (de-de)\n', () => {
+	let testCase = {
+		"und — sie sagte": "und – sie sagte",
+		};
+
+	Object.keys(testCase).forEach((key) => {
+		it("unit test", () => {
+			assert.equal(consolidateSpacedDashes(key, new Locale("de-de")), testCase[key]);
+		});
+	});
+});
+
+describe('Fix dash spaces between words (en-us)\n', () => {
 	let testCase = {
 		"and — she said": "and—she said",
 		"and — she said": "and—she said", //mixed spaces
+		"and— she said": "and—she said",
+		"and —she said": "and—she said",
 		"and—she said": "and—she said",
 		};
 
 	Object.keys(testCase).forEach((key) => {
 		it("unit test", () => {
-			assert.equal(fixSpacesAroundEmDash(key, new Locale("en-us")), testCase[key]);
+			assert.equal(fixDashSpacesBetweenWords(key, new Locale("en-us")), testCase[key]);
 		});
 	});
 });
 
-describe('Fix spaces around em dash in Rusyn, Slovak, Czech\n', () => {
+describe('Fix dash spaces between words (rue, sk)\n', () => {
 	let testCase = {
 		"and — she said": "and — she said",
 		"and—she said": "and — she said",
@@ -82,9 +94,39 @@ describe('Fix spaces around em dash in Rusyn, Slovak, Czech\n', () => {
 
 	Object.keys(testCase).forEach((key) => {
 		it("unit test", () => {
-			assert.equal(fixSpacesAroundEmDash(key, new Locale("rue")), testCase[key]);
-			assert.equal(fixSpacesAroundEmDash(key, new Locale("sk")), testCase[key]);
-			assert.equal(fixSpacesAroundEmDash(key, new Locale("cs")), testCase[key]);
+			assert.equal(fixDashSpacesBetweenWords(key, new Locale("rue")), testCase[key]);
+			assert.equal(fixDashSpacesBetweenWords(key, new Locale("sk")), testCase[key]);
+		});
+	});
+});
+
+describe('Fix dash spaces between words (cs)\n', () => {
+	let testCase = {
+		"and — she said": "and – she said",
+		"and—she said": "and – she said",
+		};
+
+	Object.keys(testCase).forEach((key) => {
+		it("unit test", () => {
+			assert.equal(fixDashSpacesBetweenWords(key, new Locale("cs")), testCase[key]);
+		});
+	});
+});
+
+describe('Fix dash spaces between words (de-de) \n', () => {
+	let testCase = {
+		"und –sie sagte": "und – sie sagte",
+		"und– sie sagte": "und – sie sagte",
+		"und – sie sagte": "und – sie sagte", //mixed spaces
+		"und–sie sagte": "und – sie sagte",
+		};
+
+	Object.keys(testCase).forEach((key) => {
+		it("unit test", () => {
+			assert.equal(fixDashSpacesBetweenWords(key, new Locale("de-de")), testCase[key]);
+		});
+		it("module test", () => {
+			assert.equal(fixDash(key, new Locale("de-de")), testCase[key]);
 		});
 	});
 });
@@ -101,6 +143,12 @@ describe('Fix dash between cardinal numbers\n', () => {
 		"In 5,000.25-10,000.75 range": "In 5,000.25–10,000.75 range",
 		"v rozmedzí 5,25-10,75": "v rozmedzí 5,25–10,75",
 		"v rozmedzí 5 000,25-10 000,75": "v rozmedzí 5 000,25–10 000,75",
+		"2-3 Eier": "2–3 Eier",
+		"2 -3 Eier": "2–3 Eier",
+		"2- 3 Eier": "2–3 Eier",
+		"2 - 3 Eier": "2–3 Eier",
+		"2—3 Eier": "2–3 Eier",
+		"im Bereich von 5.000,25-10.000,75": "im Bereich von 5.000,25–10.000,75",
 		};
 
 	Object.keys(testCase).forEach((key) => {
@@ -113,7 +161,7 @@ describe('Fix dash between cardinal numbers\n', () => {
 	});
 });
 
-describe('Fix dash between ordinal numbers in English\n', () => {
+describe('Fix dash between ordinal numbers (en-us)\n', () => {
 	let testCase = {
 		"1st-5th August": "1st–5th August",
 		"1st -5th August": "1st–5th August",
@@ -131,7 +179,7 @@ describe('Fix dash between ordinal numbers in English\n', () => {
 	});
 });
 
-describe('Fix dash between ordinal numbers in Rusyn, Slovak, Czech\n', () => {
+describe('Fix dash between ordinal numbers (rue, sk, cs, de)\n', () => {
 	let testCase = {
 		"1.-5. augusta": "1.–5. augusta",
     "1. -5. augusta": "1.–5. augusta",
@@ -144,11 +192,13 @@ describe('Fix dash between ordinal numbers in Rusyn, Slovak, Czech\n', () => {
 			assert.equal(fixDashBetweenOrdinalNumbers(key, new Locale("rue")), testCase[key]);
 			assert.equal(fixDashBetweenOrdinalNumbers(key, new Locale("sk")), testCase[key]);
 			assert.equal(fixDashBetweenOrdinalNumbers(key, new Locale("cs")), testCase[key]);
+			assert.equal(fixDashBetweenOrdinalNumbers(key, new Locale("de-de")), testCase[key]);
 		});
 		it("module test", () => {
 			assert.equal(fixDash(key, new Locale("rue")), testCase[key]);
 			assert.equal(fixDash(key, new Locale("sk")), testCase[key]);
 			assert.equal(fixDash(key, new Locale("cs")), testCase[key]);
+			assert.equal(fixDash(key, new Locale("de-de")), testCase[key]);
 		});
 	});
 });