+++ src/test/whitespace/nbsp.test.js
@@ -6,10 +6,8 @@ import {removeNbspBetweenMultiCharWords,
 				addNbspAfterOrdinalNumber,
 				addNbspWithinOrdinalDate,
 				addNbspAfterRomanNumeral,
-				addNbspAfterInitial,
 				addNbspAfterSymbol,
 				replaceSpacesWithNbspAfterSymbol,
-				addNbspAfterAbbreviation,
 				fixNbsp} from "../../lib/whitespace/nbsp";
 import assert from 'assert';
 import Locale from "../../locale/locale";
@@ -34,6 +32,8 @@ describe('Remove non-breaking space between multi-letter words\n', () => {
 		});
 	});
 });
+
+
 
 describe('Add non-breaking spaces after single-character prepositions\n', () => {
 	let testCase = {
@@ -110,7 +110,7 @@ describe('Add non-breaking space after ordinal number (en)\n', () => {
 		it("unit test", () => {
 			assert.equal(addNbspAfterOrdinalNumber(key, new Locale("en-us")), testCase[key]);
 		});
-		it("moduel test", () => {
+		it("module test", () => {
 			assert.equal(fixNbsp(key, new Locale("en-us")), testCase[key]);
 		});
 	});
@@ -118,7 +118,7 @@ describe('Add non-breaking space after ordinal number (en)\n', () => {
 
 
 
-describe('Add non-breaking space after ordinal number (sk, cs, rue)\n', () => {
+describe('Add non-breaking space after ordinal number (sk, cs, rue, de-de)\n', () => {
 	let testCase = {
 		"1. dodatok" : "1. dodatok",
 		"1.dodatok" : "1. dodatok",
@@ -132,11 +132,13 @@ describe('Add non-breaking space after ordinal number (sk, cs, rue)\n', () => {
 		it("unit test", () => {
 			assert.equal(addNbspAfterOrdinalNumber(key, new Locale("sk")), testCase[key]);
 			assert.equal(addNbspAfterOrdinalNumber(key, new Locale("cs")), testCase[key]);
+			assert.equal(addNbspAfterOrdinalNumber(key, new Locale("de-de")), testCase[key]);
 			assert.equal(addNbspAfterOrdinalNumber(key, new Locale("rue")), testCase[key]);
 		});
 		it("module test", () => {
 			assert.equal(fixNbsp(key, new Locale("sk")), testCase[key]);
 			assert.equal(fixNbsp(key, new Locale("cs")), testCase[key]);
+			assert.equal(fixNbsp(key, new Locale("de-de")), testCase[key]);
 			assert.equal(fixNbsp(key, new Locale("rue")), testCase[key]);
 		});
 	});
@@ -167,8 +169,32 @@ describe('Add non-breaking space within ordinal date (sk, cs, rue)\n', () => {
 
 
 
-describe('Add non-breaking space after roman numeral (sk, cs, rue)\n', () => {
+describe('Add non-breaking space within ordinal date (de-de)\n', () => {
 	let testCase = {
+		/*  German standard orthography (Duden) recommends
+				only one nbsp (or narrowNbsp) after the day
+				and a regular interword space following the month*/
+		"12.1.2019" : "12. 1. 2019",
+		"12. 1. 2019": "12. 1. 2019",
+		"10.00" : "10.00", // false positive for the example above
+	};
+
+	Object.keys(testCase).forEach((key) => {
+		it("unit test", () => {
+			assert.equal(addNbspWithinOrdinalDate(key, new Locale("de-de")), testCase[key]);
+		});
+		it("module test", () => {
+			assert.equal(fixNbsp(key, new Locale("de-de")), testCase[key]);
+		});
+	});
+});
+
+
+
+describe('Add non-breaking space after roman numeral (sk, cs, de-de, rue)\n', () => {
+	let testCase = {
+		"I. kapitola": "I. kapitola",
+		"bola to I. kapitola": "bola to I. kapitola",
 		"III. kapitola": "III. kapitola",
 		"III.kapitola": "III. kapitola",
 		"X. ročník": "X. ročník",
@@ -183,30 +209,14 @@ describe('Add non-breaking space after roman numeral (sk, cs, rue)\n', () => {
 		it("unit test", () => {
 			assert.equal(addNbspAfterRomanNumeral(key, new Locale("sk")), testCase[key]);
 			assert.equal(addNbspAfterRomanNumeral(key, new Locale("cs")), testCase[key]);
+			assert.equal(addNbspAfterRomanNumeral(key, new Locale("de-de")), testCase[key]);
 			assert.equal(addNbspAfterRomanNumeral(key, new Locale("rue")), testCase[key]);
 		});
 		it("module test", () => {
 			assert.equal(fixNbsp(key, new Locale("sk")), testCase[key]);
 			assert.equal(fixNbsp(key, new Locale("cs")), testCase[key]);
+			assert.equal(fixNbsp(key, new Locale("de-de")), testCase[key]);
 			assert.equal(fixNbsp(key, new Locale("rue")), testCase[key]);
-		});
-	});
-});
-
-
-
-describe('Add non-breaking space after initial\n', () => {
-	let testCase = {
-		"Philip K. Dick": "Philip K. Dick",
-		"F. X. Šalda": "F. X. Šalda",
-	};
-
-	Object.keys(testCase).forEach((key) => {
-		it("unit test", () => {
-			assert.equal(addNbspAfterInitial(key, new Locale("en-us")), testCase[key]);
-		});
-		it("module test", () => {
-			assert.equal(fixNbsp(key, new Locale("en-us")), testCase[key]);
 		});
 	});
 });
@@ -238,43 +248,6 @@ describe('Replaces various spaces with non-breaking space after symbol, e.g. ©\
 	Object.keys(testCase).forEach((key) => {
 		it("just unit tests", () => {
 			assert.equal(replaceSpacesWithNbspAfterSymbol(key, new Locale("en-us"), "©"), testCase[key]);
-		});
-	});
-});
-
-
-
-describe('Add non-breaking space after abbreviations\n', () => {
-	let testCase = {
-		// as all abbreviations can be used within different languages, such as reference, they're being corrected in each language
-		"č.5 žije" : "č. 5 žije",
-		"Č.5 žije" : "č. 5 žije",
-		"č. 5 žije" : "č. 5 žije",
-		"áno, č. 5 žije" : "áno, č. 5 žije",
-		"preč č. 5 žije" : "preč č. 5 žije",
-		"str. 8" : "str. 8",
-		"tzv. rýč" : "tzv. rýč",
-		"Prines kvetináč. 5 je super číslo." : "Prines kvetináč. 5 je super číslo.", //false positive
-		"hl. m. Praha" : "hl. m. Praha",
-		// "hl.m.Praha" : "hl. m. Praha", // 2 consequetive abbreviations are not supported yet
-		// "hl. m.Praha" : "hl. m. Praha",
-		// "hl.m. Praha" : "hl. m. Praha",
-		"10 a.m." : "10 a.m.", // false positive for abbreviation within abbreviation, i.e. m. within a.m.
-		"e.g. something" : "e.g. something",
-		"10 p." : "10 p.",
-		"pp. 10–25" : "pp. 10–25",
-		"(e.g.)" : "(e.g.)",
-		"“e.g.”" : "“e.g.”",
-		"‘e.g.’" : "‘e.g.’",
-		"John Thune (S.D.)" : "John Thune (S.D.)", // false positive
-
-
-
-	};
-
-	Object.keys(testCase).forEach((key) => {
-		it("unit tests", () => {
-			assert.equal(addNbspAfterAbbreviation(key, new Locale("en-us")), testCase[key]);
 		});
 	});
 });